Several Changes
Changes how the connection to the vera starts. micasaverde_vera.connect(ip)

Removes the get_device, get_scene.... functions in favor of __getattr__ and __getitem__. one can now do vera.devices[32]  or vera.devices['Test Device']. This will also work as well. vera.devices.Kitchen for single word names not containing any special characters. otherwise you can use the previous method or getattr(vera.devices, 'Master Bedroom')

Changes the whole event system to be uniform with the class instance structure.
    as an example. an event founs  be "devices.1.name.changed'. this event naming system directly matches the organization layout. so vera.devices[1].name.

Adds automatic grabbing of categories and sub categories. and if there is no category or subcategory variables associated with a device it will check to see if it is a plugin device. and if so it will return the plugin name

also added the returnning of a plugin instance if the plugin variable is available for a device. so if one did. device.plugin it would not be a number but a plugin instance.

changed the class structure for scenes. they now become subclasses of SceneController1 and Scene1 respectivly. the scenes class can be accessed one of 2 ways. vera.scenes or also as a device, vera.devices['_Scene Controller']. This was done to inherit all of the available functions that get generated by the build process.