Adds dynamic addition of classes not listed in device files
There are cases where a service is not listed in the device files. This will dynamically add the class to the device class to populate any properties and functions into that device. there is a service named ZWaveDevice1 that there is no service file for and would cause a rebuilding loop. I manually typed up this service class and it will dynamically add the service if needed