Adds device command documentation
This is a simple breakdown of all the stock devices that are available on the Vera.

For each available device type it lists off the services associated with that device along with the variables associated to that service as well as any functions along with the arguments for that function. It also gives an example of the command that is sent to the Vera for that specific item.