Adds threaded events
I have added the creation of a thread or each event that gets generated.
This is done so the user has the ability to run whatever code they wish
without stalling the update process.

Also fixes some issues with the scenes not having proper parameters.
A lot of code cleanup.

Changes version so the generated files will automatically be rebuilt.